Linux 2.4.0-test9pre5
    - Make SCSI initialization order be same as before.
    - fix cardbus bridge resources..
    - don't disallow Onstream ide-scsi devices
    - byteorder: use statement expressions instead of macros, to avoid argument re-use.
    - codafs update
    - more USB updates
    - _fput/__fput are no longer used.
    - ixj telephony driver fixes
    - pmac SCSI driver init update
    - Andries: net device name allocation as in 2.2.x
    - sis900 driver update
    - more drivers synced to Alan's 2.2.x changes

+++ drivers/char/drm/r128_drv.h
@@ -757,7 +757,7 @@ macro_line|#       define R128_CCE_PACKET1_REG1_MASK&t;0x003ff800
 DECL|macro|R128_MAX_USEC_TIMEOUT
 mdefine_line|#define R128_MAX_USEC_TIMEOUT&t;100000&t;/* 100 ms */
 DECL|macro|R128_BASE
-mdefine_line|#define R128_BASE(reg)&t;&t;((u32)(dev_priv-&gt;mmio-&gt;handle))
+mdefine_line|#define R128_BASE(reg)&t;&t;((unsigned long)(dev_priv-&gt;mmio-&gt;handle))
 DECL|macro|R128_ADDR
 mdefine_line|#define R128_ADDR(reg)&t;&t;(R128_BASE(reg) + reg)
 DECL|macro|R128_DEREF
